feat: add custom sound player support
This commit introduces the ability to specify a custom program to play
the alarm sound at the end of a pomodoro session. The program can be
specified via the -p command line option or the POMODORO_SOUND_PLAYER
environment variable. If no program is specified, 'mplayer' is used by
default. The commit also updates the help message to reflect these
changes.